Add catalog health backoffice
