+++ .github/workflows/container-images.yml
@@ -30,6 +30,8 @@ jobs:
             dockerfile: apps/web/workers.Dockerfile
           - name: bull-board
             dockerfile: apps/web/bull-board.Dockerfile
+          - name: backoffice
+            dockerfile: apps/backoffice/Dockerfile
           - name: docs
             dockerfile: apps/docs/Dockerfile
           - name: db-migrate
