+++ README.md
@@ -222,6 +222,7 @@ renders the same files from `apps/docs`:
 ```bash
 npm run dev:docs     # http://localhost:3003
 npm run build:docs   # production docs build
+npm run dev:backoffice # catalog-health operator UI
 ```
 
 - **[Setup Guide](./docs/SETUP.md)** — Complete setup instructions
@@ -277,8 +278,10 @@ For ownership rules inside `apps/web/src`, see [docs/BOUNDARIES.md](./docs/BOUND
 # Development
 npm run dev                         # Start development server (repo root)
 npm run dev:docs                    # Start documentation site at http://localhost:3003
+npm run dev:backoffice              # Start catalog-health operator UI
 npm run build                       # Build for production (repo root)
 npm run build:docs                  # Build the documentation site
+npm run build:backoffice            # Build the backoffice app
 npm run start --workspace=apps/web  # Start production server (apps/web)
 cd apps/web && npm run start:workers # Start BullMQ workers (apps/web)
 cd apps/web && npm run bull-board    # Launch BullMQ dashboard (apps/web)
