+++ apps/backoffice/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@pop-choice/backoffice",
+  "version": "0.1.0-beta.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsx --env-file=.env src/index.ts",
+    "start": "node dist/index.js",
+    "type-check": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@pop-choice/shared": "*",
+    "express": "^5.2.1",
+    "express-rate-limit": "^8.5.2"
+  },
+  "devDependencies": {
+    "@types/express": "^5.0.6",
+    "@types/node": "25.9.1",
+    "tsx": "^4.22.3",
+    "typescript": "6.0.3"
+  },
+  "engines": {
+    "node": ">=24",
+    "npm": ">=11"
+  }
+}
