+++ coolify.compose.yml
@@ -164,6 +164,36 @@ services:
       retries: 5
       start_period: 30s
 
+  backoffice:
+    image: ${APP_IMAGE_PREFIX:-ghcr.io/shchilkin/popchoice}/backoffice:${IMAGE_TAG:-development}
+    pull_policy: always
+    restart: unless-stopped
+    command: ['npm', 'run', 'start', '--workspace=apps/backoffice']
+    environment:
+      <<: *app-env
+      PORT: 3000
+      OPERATOR_AUTH_REQUIRED: ${OPERATOR_AUTH_REQUIRED:-0}
+      OPERATOR_AUTH_USERNAME: ${OPERATOR_AUTH_USERNAME:-}
+      OPERATOR_AUTH_PASSWORD: ${OPERATOR_AUTH_PASSWORD:-}
+      OPERATOR_AUTH_REALM: ${OPERATOR_AUTH_REALM:-PopChoice Operators}
+      CATALOG_HEALTH_SAMPLE_LIMIT: ${CATALOG_HEALTH_SAMPLE_LIMIT:-5}
+      CATALOG_HEALTH_STALE_DAYS: ${CATALOG_HEALTH_STALE_DAYS:-180}
+    depends_on:
+      db:
+        condition: service_healthy
+    expose:
+      - '3000'
+    healthcheck:
+      test:
+        [
+          'CMD-SHELL',
+          'node -e "fetch(''http://127.0.0.1:3000/healthz'').then(r=>process.exit(r.ok?0:1)).catch(()=>process.exit(1))"',
+        ]
+      interval: 30s
+      timeout: 10s
+      retries: 5
+      start_period: 30s
+
   docs:
     image: ${APP_IMAGE_PREFIX:-ghcr.io/shchilkin/popchoice}/docs:${IMAGE_TAG:-development}
     pull_policy: always
