+++ docs/ROADMAP-ARCHITECTURE.md
@@ -193,7 +193,7 @@ This is an extraction direction, not a mandate for large-scale file moves right 
 - [x] [#472](https://github.com/shchilkin/PopChoice/issues/472): extend TMDB backfill/discovery to populate people, genre, and keyword metadata, with catalog-health visibility for missing coverage.
 - [x] Split the backoffice/catalog-health epic [#493](https://github.com/shchilkin/PopChoice/issues/493) into implementation-sized child issues before broad implementation starts.
 - [x] Define the dedicated backoffice app boundaries, routing, deployment, and ownership model in [#547](https://github.com/shchilkin/PopChoice/issues/547). The backoffice is planned as `apps/backoffice`, deployed like `apps/bull-board`, not as UI inside `apps/web`.
-- Build the protected catalog-health overview in [#549](https://github.com/shchilkin/PopChoice/issues/549), covering duplicate movie identities, missing posters, missing localized names/overviews, missing runtimes, stale TMDB metadata, and missing cast/director/genre/keyword coverage.
+- [x] Build the protected catalog-health overview in [#549](https://github.com/shchilkin/PopChoice/issues/549), covering duplicate movie identities, missing posters, missing localized names/overviews, missing runtimes, stale TMDB metadata, and missing cast/director/genre/keyword coverage.
 - Persist TMDB backfill review cases in `tmdb_match_reviews`, then expose a protected TMDB match review queue in [#550](https://github.com/shchilkin/PopChoice/issues/550) for ambiguous matches, missing metadata, and rejected runtime/year confidence cases.
 - Add safe review actions and audit/history rules in [#551](https://github.com/shchilkin/PopChoice/issues/551) before allowing operators to apply, reject, or defer catalog fixes.
 - Periodically refresh TMDB-backed metadata for older records without destabilizing existing recommendation history.
